refactor(anchor-navigation-item): convert to typescript
